Extract azure sdk project dependency panel
