+++ PluginsAndFeatures/azure-toolkit-for-intellij/azure-sdk-reference-book/src/main/java/com/microsoft/azure/toolkit/intellij/azuresdk/referencebook/components/ModuleComboBox.java
@@ -7,6 +7,7 @@ package com.microsoft.azure.toolkit.intellij.azuresdk.referencebook.components;
 
 import com.intellij.openapi.project.Project;
 import com.intellij.openapi.project.ProjectManager;
+import com.intellij.ui.components.fields.ExtendableTextComponent;
 import com.microsoft.azure.toolkit.intellij.azuresdk.model.module.GradleProjectModule;
 import com.microsoft.azure.toolkit.intellij.azuresdk.model.module.MavenProjectModule;
 import com.microsoft.azure.toolkit.intellij.azuresdk.model.module.ProjectModule;
@@ -16,6 +17,7 @@ import org.apache.commons.lang3.StringUtils;
 import javax.annotation.Nonnull;
 import javax.annotation.Nullable;
 import javax.swing.*;
+import java.util.Collections;
 import java.util.Comparator;
 import java.util.List;
 import java.util.stream.Collectors;
@@ -41,6 +43,12 @@ public class ModuleComboBox extends AzureComboBox<ProjectModule> {
         return Stream.of(mavenProjectModules, gradleProjectModules).flatMap(List::stream)
                 .sorted((Comparator<ProjectModule>) (first, second) -> StringUtils.compare(first.getName(), second.getName()))
                 .collect(Collectors.toList());
+    }
+
+    @Nonnull
+    @Override
+    protected List<ExtendableTextComponent.Extension> getExtensions() {
+        return Collections.emptyList();
     }
 
     @Override
